requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
dnspython==2.4.2
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mongomock==4.1.2
motor==3.3.2
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.6.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
python-on-whales==0.67.0
PyYAML==6.0.1
requests==2.31.0
sentinels==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.37
starlette==0.27.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.3
